fix missing fields issue in auth flow by creating a separate `withIsLoggedIn` HOC to track
