+++ src/@ui/NativeWebRouter/ProtectedRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { compose, branch, renderComponent } from 'recompose';
-import withUser from '@data/withUser';
+import withUser from '@data/withUser/withIsLoggedIn';
 import { Route, Redirect } from 'react-router';
 import ActivityIndicator from '@ui/ActivityIndicator';
 
@@ -30,10 +30,10 @@ ActivityIndicatorWhileLoading.propTypes = { ...Route.propTypes };
 
 const ProtectedRoute = compose(
   withUser,
-  branch(({ isLoading, user }) => isLoading && !user,
+  branch(({ isLoading }) => isLoading,
     renderComponent(ActivityIndicatorWhileLoading),
   ),
-  branch(({ user }) => !user,
+  branch(({ isLoggedIn }) => !isLoggedIn,
     renderComponent(RedirectWithReferrer),
   ),
 )(Route);
